Correctly parse translations for blank lyrics
Some lyrics providers have lyrics that use duplicated timestamps for formatting reasons, causing the blank line to be parsed as the main lyric, and the second (actual line) to be parsed as a translation. After this change, blank lines will not be eligible for translation.

See the following for an example:

[00:23.73]I see the world through eyes covered in ink and bleach
[00:29.70]Cross out the ones who heard my cries and watched me weep
[00:35.64]
[00:35.64]I love everything
[00:38.51]Fire's spreading all around my room
[00:43.13]My world's so bright
[00:44.42]It's hard to breathe but that's alright
[00:47.32]Hush
[00:59.22]
[00:59.22]Shh
[01:11.46]
[01:11.46]Tape my eyes open to force reality (oh no, no)
[01:17.99]Why can't you just let me eat my weight in glee?